test: migrated federation-with-gql.test.js and federation.test.js from tap to node.test
