+++ test/federation-with-gql.test.js
@@ -1,24 +1,14 @@
 'use strict'
 
-const { test } = require('tap')
+const { test } = require('node:test')
 const Fastify = require('fastify')
-const { printSchema, defaultFieldResolver } = require('graphql')
+const { printSchema } = require('graphql')
 const WebSocket = require('ws')
 const mq = require('mqemitter')
-const GQL = require('mercurius')
-const mercuriusWithFederation = require('../lib/plugin')
-
-const { makeExecutableSchema } = require('@graphql-tools/schema')
-const { mergeResolvers } = require('@graphql-tools/merge')
-const {
-  MapperKind,
-  mapSchema,
-  getDirectives,
-  printSchemaWithDirectives,
-  getResolversFromSchema
-} = require('@graphql-tools/utils')
-
+const mercurius = require('mercurius')
+const gql = require('graphql-tag')
 const buildFederationSchema = require('../lib/federation')
+const { matchSnapshot } = require('./helpers/snapshotHelper')
 
 test('federation support using schema from buildFederationSchema', async t => {
   const app = Fastify()
@@ -37,7 +27,6 @@ test('federation support using schema from buildFederationSchema', async t => {
       username: String
     }
   `
-
   const resolvers = {
     Query: {
       me: () => ({
@@ -60,9 +49,9 @@ test('federation support using schema from buildFederationSchema', async t => {
     }
   }
 
-  const federationSchema = buildFederationSchema(schema)
+  const federationSchema = buildFederationSchema(gql(schema))
 
-  app.register(GQL, {
+  app.register(mercurius, {
     schema: federationSchema,
     resolvers
   })
@@ -71,17 +60,18 @@ test('federation support using schema from buildFederationSchema', async t => {
 
   let query = '{ _service { sdl } }'
   let res = await app.inject({ method: 'GET', url: `/graphql?query=${query}` })
-  t.same(JSON.parse(res.body), { data: { _service: { sdl: schema } } })
+
+  t.assert.deepStrictEqual(JSON.parse(res.body), { data: { _service: { sdl: schema } } })
 
   query = '{ me { id name username } }'
   res = await app.inject({ method: 'GET', url: `/graphql?query=${query}` })
-  t.same(JSON.parse(res.body), {
+  t.assert.deepStrictEqual(JSON.parse(res.body), {
     data: { me: { id: '1', name: 'John', username: '@john' } }
   })
 
   query = 'mutation { add(a: 11 b: 19) }'
   res = await app.inject({ method: 'POST', url: '/graphql', body: { query } })
-  t.same(JSON.parse(res.body), { data: { add: 30 } })
+  t.assert.deepStrictEqual(JSON.parse(res.body), { data: { add: 30 } })
 })
 
 test('a normal schema can be run in federated mode', async t => {
@@ -110,9 +100,9 @@ test('a normal schema can be run in federated mode', async t => {
     }
   }
 
-  const federationSchema = buildFederationSchema(schema)
+  const federationSchema = buildFederationSchema(gql(schema))
 
-  app.register(GQL, {
+  app.register(mercurius, {
     schema: federationSchema,
     resolvers
   })
@@ -125,7 +115,7 @@ test('a normal schema can be run in federated mode', async t => {
     url: `/graphql?query=${query}`
   })
 
-  t.same(JSON.parse(res.body), {
+  t.assert.deepStrictEqual(JSON.parse(res.body), {
     data: {
       _service: {
         sdl: schema
@@ -156,9 +146,9 @@ test('a schema can be run in federated mode when Query is not defined', async t 
     }
   }
 
-  const federationSchema = buildFederationSchema(schema)
+  const federationSchema = buildFederationSchema(gql(schema))
 
-  app.register(GQL, {
+  app.register(mercurius, {
     schema: federationSchema,
     resolvers
   })
@@ -171,7 +161,7 @@ test('a schema can be run in federated mode when Query is not defined', async t 
     url: `/graphql?query=${query}`
   })
 
-  t.same(JSON.parse(res.body), {
+  t.assert.deepStrictEqual(JSON.parse(res.body), {
     data: {
       _service: {
         sdl: schema
@@ -215,9 +205,9 @@ test('entities resolver returns correct value', async t => {
     }
   }
 
-  const federationSchema = buildFederationSchema(schema)
+  const federationSchema = buildFederationSchema(gql(schema))
 
-  app.register(GQL, {
+  app.register(mercurius, {
     schema: federationSchema,
     resolvers
   })
@@ -241,7 +231,7 @@ test('entities resolver returns correct value', async t => {
     url: `/graphql?query=${query}`
   })
 
-  t.same(JSON.parse(res.body), {
+  t.assert.deepStrictEqual(JSON.parse(res.body), {
     data: {
       _entities: [
         {
@@ -290,9 +280,9 @@ test('entities resolver returns correct value with async resolver', async t => {
     }
   }
 
-  const federationSchema = buildFederationSchema(schema)
+  const federationSchema = buildFederationSchema(gql(schema))
 
-  app.register(GQL, {
+  app.register(mercurius, {
     schema: federationSchema,
     resolvers
   })
@@ -316,7 +306,7 @@ test('entities resolver returns correct value with async resolver', async t => {
     url: `/graphql?query=${query}`
   })
 
-  t.same(JSON.parse(res.body), {
+  t.assert.deepStrictEqual(JSON.parse(res.body), {
     data: {
       _entities: [
         {
@@ -356,9 +346,9 @@ test('entities resolver returns user default resolver if resolveReference is not
     }
   }
 
-  const federationSchema = buildFederationSchema(schema)
+  const federationSchema = buildFederationSchema(gql(schema))
 
-  app.register(GQL, {
+  app.register(mercurius, {
     schema: federationSchema,
     resolvers
   })
@@ -382,7 +372,7 @@ test('entities resolver returns user default resolver if resolveReference is not
     url: `/graphql?query=${query}`
   })
 
-  t.same(JSON.parse(res.body), {
+  t.assert.deepStrictEqual(JSON.parse(res.body), {
     data: {
       _entities: [
         {
@@ -422,9 +412,9 @@ test('entities resolver throws an error if reference type name not in schema', a
     }
   }
 
-  const federationSchema = buildFederationSchema(schema)
+  const federationSchema = buildFederationSchema(gql(schema))
 
-  app.register(GQL, {
+  app.register(mercurius, {
     schema: federationSchema,
     resolvers
   })
@@ -448,7 +438,7 @@ test('entities resolver throws an error if reference type name not in schema', a
     url: `/graphql?query=${query}`
   })
 
-  t.equal(
+  t.assert.strictEqual(
     JSON.parse(res.body).errors[0].message,
     'The _entities resolver tried to load an entity for type "Account", but no object type of that name was found in the schema'
   )
@@ -473,9 +463,9 @@ test('buildFederationSchema function adds stub types', async t => {
     directive @customdir on FIELD_DEFINITION
   `
 
-  const federationSchema = buildFederationSchema(schema)
+  const federationSchema = buildFederationSchema(gql(schema))
 
-  t.matchSnapshot(printSchema(federationSchema))
+  matchSnapshot(printSchema(federationSchema), { name: 'federation-with-gql' })
 })
 
 test('buildFederationSchema works correctly with multiple type extensions', async t => {
@@ -501,10 +491,10 @@ test('buildFederationSchema works correctly with multiple type extensions', asyn
     }
   `
   try {
-    buildFederationSchema(schema)
-    t.pass('schema built without errors')
+    buildFederationSchema(gql(schema))
+    t.assert.ok('schema built without errors')
   } catch (err) {
-    t.fail('it should not throw errors', err)
+    t.assert.fail(['it should not throw errors', err])
   }
 })
 
@@ -535,10 +525,10 @@ test('buildFederationSchema ignores UniqueDirectivesPerLocationRule when validat
     }
   `
   try {
-    buildFederationSchema(schema)
-    t.pass('schema built without errors')
+    buildFederationSchema(gql(schema))
+    t.assert.ok('schema built without errors')
   } catch (err) {
-    t.fail('it should not throw errors', err)
+    t.assert.fail(['it should not throw errors', err])
   }
 })
 
@@ -565,11 +555,11 @@ test('buildFederationSchema still validate schema for errors (1 error)', async t
     }
   `
   try {
-    buildFederationSchema(schema)
-    t.fail('it should throw validation error')
+    buildFederationSchema(gql(schema))
+    t.assert.fail('it should throw validation error')
   } catch (err) {
     // expected error: Field "User.id" can only be defined once.
-    t.pass('it should throw error')
+    t.assert.ok('it should throw error')
   }
 })
 
@@ -597,15 +587,15 @@ test('buildFederationSchema still validate schema for errors (multiple error)', 
     }
   `
   try {
-    buildFederationSchema(schema)
-    t.fail('it should throw validation error')
+    buildFederationSchema(gql(schema))
+    t.assert.fail('it should throw validation error')
   } catch (err) {
     // expected errors:
     // Field "User.id" can only be defined once.
     // Unknown type "Actions"
-    t.ok(err.errors)
-    t.equal(err.errors.length, 2)
-    t.pass('it should throw error')
+    t.assert.ok(err.errors)
+    t.assert.strictEqual(err.errors.length, 2)
+    t.assert.ok('it should throw error')
   }
 })
 
@@ -653,9 +643,9 @@ test('mutation works with federation support', async t => {
     }
   }
 
-  const federationSchema = buildFederationSchema(schema)
+  const federationSchema = buildFederationSchema(gql(schema))
 
-  app.register(GQL, {
+  app.register(mercurius, {
     schema: federationSchema,
     resolvers
   })
@@ -671,7 +661,7 @@ test('mutation works with federation support', async t => {
     }
   })
 
-  t.same(JSON.parse(res.body), {
+  t.assert.deepStrictEqual(JSON.parse(res.body), {
     data: {
       add: 30
     }
@@ -680,7 +670,7 @@ test('mutation works with federation support', async t => {
 
 test('subscription server sends update to subscriptions', t => {
   const app = Fastify()
-  t.teardown(() => app.close())
+  t.after(() => app.close())
 
   const sendTestQuery = () => {
     app.inject(
@@ -781,9 +771,9 @@ test('subscription server sends update to subscriptions', t => {
     }
   }
 
-  const federationSchema = buildFederationSchema(schema)
+  const federationSchema = buildFederationSchema(gql(schema))
 
-  app.register(GQL, {
+  app.register(mercurius, {
     schema: federationSchema,
     resolvers,
     subscription: {
@@ -792,7 +782,7 @@ test('subscription server sends update to subscriptions', t => {
   })
 
   app.listen({ port: 0 }, err => {
-    t.error(err)
+    t.assert.ifError(err)
 
     const ws = new WebSocket(
       'ws://localhost:' + app.server.address().port + '/graphql',
@@ -802,7 +792,7 @@ test('subscription server sends update to subscriptions', t => {
       encoding: 'utf8',
       objectMode: true
     })
-    t.teardown(client.destroy.bind(client))
+    t.after(() => client.destroy.bind(client))
     client.setEncoding('utf8')
 
     client.write(
@@ -873,7 +863,6 @@ test('subscription server sends update to subscriptions', t => {
         )
 
         client.end()
-        t.end()
       } else if (data.id === 2 && data.type === 'complete') {
         sendTestQuery()
       }
@@ -918,7 +907,7 @@ test('federation supports loader for __resolveReference function', async t => {
   const loaders = {
     User: {
       async __resolveReference (queries) {
-        t.same(queries, [
+        t.assert.deepEqual(queries, [
           {
             obj: {
               __typename: 'User',
@@ -939,9 +928,9 @@ test('federation supports loader for __resolveReference function', async t => {
     }
   }
 
-  const federationSchema = buildFederationSchema(schema)
+  const federationSchema = buildFederationSchema(gql(schema))
 
-  app.register(GQL, {
+  app.register(mercurius, {
     schema: federationSchema,
     resolvers,
     loaders
@@ -968,7 +957,7 @@ test('federation supports loader for __resolveReference function', async t => {
     }
   })
 
-  t.same(JSON.parse(res.body), {
+  t.assert.deepStrictEqual(JSON.parse(res.body), {
     data: {
       _entities: [
         {
@@ -1011,9 +1000,9 @@ test('federation schema is built correctly with type extension', async t => {
     }
   `
 
-  const federationSchema = buildFederationSchema(schema)
+  const federationSchema = buildFederationSchema(gql(schema))
 
-  app.register(GQL, {
+  app.register(mercurius, {
     schema: federationSchema
   })
 
@@ -1025,7 +1014,7 @@ test('federation schema is built correctly with type extension', async t => {
     url: `/graphql?query=${query}`
   })
 
-  t.same(JSON.parse(res.body), {
+  t.assert.deepStrictEqual(JSON.parse(res.body), {
     data: {
       topPosts: null
     }
@@ -1071,9 +1060,9 @@ test("basic federation support with 'schema' in the schema", async t => {
     }
   }
 
-  const federationSchema = buildFederationSchema(schema)
+  const federationSchema = buildFederationSchema(gql(schema))
 
-  app.register(GQL, {
+  app.register(mercurius, {
     schema: federationSchema,
     resolvers
   })
@@ -1086,7 +1075,7 @@ test("basic federation support with 'schema' in the schema", async t => {
     url: `/graphql?query=${query}`
   })
 
-  t.same(JSON.parse(res.body), {
+  t.assert.deepStrictEqual(JSON.parse(res.body), {
     data: {
       _service: {
         sdl: schema
@@ -1122,34 +1111,36 @@ test('buildFederationSchema remove some directive in isGateway mode', async t =>
     }
   `
   try {
-    const resultGateway = buildFederationSchema(schema, { isGateway: true })
+    const resultGateway = buildFederationSchema(gql(schema), { isGateway: true })
 
     const printedSchema = printSchema(resultGateway)
 
-    t.notOk(printedSchema.includes('directive @external on FIELD_DEFINITION'))
-    t.notOk(
-      printedSchema.includes(
+    t.assert.ok(!printedSchema.includes('directive @external on FIELD_DEFINITION'))
+    t.assert.ok(
+      !printedSchema.includes(
         'directive @provides(fields: _FieldSet!) on FIELD_DEFINITION'
       )
     )
-    t.notOk(
-      printedSchema.includes(
+    t.assert.ok(
+      !printedSchema.includes(
         'directive @key(fields: _FieldSet!) on OBJECT | INTERFACE'
       )
     )
-    t.notOk(printedSchema.includes('directive @extends on OBJECT | INTERFACE'))
-    t.notOk(printedSchema.includes('type _Service'))
-    t.notOk(
-      printedSchema.includes('_entities(representations: [_Any!]!): [_Entity]!')
+    t.assert.ok(
+      !printedSchema.includes('directive @extends on OBJECT | INTERFACE')
     )
-    t.notOk(printedSchema.includes('union _Entity = User | Post'))
-    t.ok(
+    t.assert.ok(!printedSchema.includes('type _Service'))
+    t.assert.ok(
+      !printedSchema.includes('_entities(representations: [_Any!]!): [_Entity]!')
+    )
+    t.assert.ok(!printedSchema.includes('union _Entity = User | Post'))
+    t.assert.ok(
       printedSchema.includes(
         'directive @requires(fields: _FieldSet!) on FIELD_DEFINITION'
       )
     )
   } catch (err) {
-    t.fail('it should not throw errors', err)
+    t.assert.fail(['it should not throw errors', err])
   }
 })
 
@@ -1180,10 +1171,10 @@ test('buildFederationSchema thrown an error on duplicated directive that with di
     }
   `
   try {
-    buildFederationSchema(schema, { isGateway: true })
-    t.fail('schema built without errors')
+    buildFederationSchema(gql(schema), { isGateway: true })
+    t.assert.fail('schema built without errors')
   } catch (err) {
-    t.same(
+    t.assert.deepStrictEqual(
       err.message,
       'Directive with a different definition but the same name "custom" already exists in the gateway schema'
     )
@@ -1199,112 +1190,12 @@ test('buildFederationSchema with extension directive', async t => {
     }
   `
   try {
-    const resultGateway = buildFederationSchema(schema)
+    const resultGateway = buildFederationSchema(gql(schema))
     const printedSchema = printSchema(resultGateway)
-    t.ok(printedSchema.includes('directive @extends on OBJECT | INTERFACE'))
+    t.assert.ok(printedSchema.includes('directive @extends on OBJECT | INTERFACE'))
   } catch (err) {
-    t.fail('schema built with errors')
+    t.assert.fail('schema built with errors')
   }
-})
-
-test('should support directives import syntax', async (t) => {
-  const app = Fastify()
-
-  const schema = `
-    extend schema
-      @link(url: "https://specs.apollo.dev/federation/v2.0",
-        import: ["@key", "@shareable", "@override"])
-    extend type Query {
-      hello: String
-    }
-  `
-
-  const resolvers = {
-    Query: {
-      hello: () => 'world'
-    }
-  }
-  app.register(mercuriusWithFederation, {
-    schema,
-    resolvers
-  })
-
-  await app.ready()
-
-  const query = '{ _service { sdl } }'
-  const res = await app.inject({
-    method: 'GET',
-    url: `/graphql?query=${query}`
-  })
-
-  t.same(JSON.parse(res.body), {
-    data: {
-      _service: {
-        sdl: schema
-      }
-    }
-  })
-})
-
-const upperDirectiveTypeDefs = 'directive @upper on FIELD_DEFINITION'
-function upperDirectiveTransformer (schema) {
-  return mapSchema(schema, {
-    [MapperKind.OBJECT_FIELD]: (fieldConfig) => {
-      const directives = getDirectives(schema, fieldConfig)
-      for (const directive of directives) {
-        if (directive.name === 'upper') {
-          const { resolve = defaultFieldResolver } = fieldConfig
-          fieldConfig.resolve = async function (source, args, context, info) {
-            const result = await resolve(source, args, context, info)
-            if (typeof result === 'string') {
-              return result.toUpperCase()
-            }
-            return result
-          }
-          return fieldConfig
-        }
-      }
-    }
-  })
-}
-
-test('federation support using schema from buildFederationSchema and custom directives', async (t) => {
-  const app = Fastify()
-  const schema = `
-    ${upperDirectiveTypeDefs}
-    
-    type Query {
-      foo: String @upper
-    }
-  `
-
-  const resolvers = {
-    Query: {
-      foo: () => 'bar'
-    }
-  }
-
-  const federationSchema = buildFederationSchema(schema)
-
-  const executableSchema = makeExecutableSchema({
-    typeDefs: printSchemaWithDirectives(federationSchema),
-    resolvers: mergeResolvers([getResolversFromSchema(federationSchema), resolvers])
-  })
-
-  app.register(GQL, {
-    schema: executableSchema,
-    schemaTransforms: [upperDirectiveTransformer]
-  })
-
-  await app.ready()
-
-  let query = '{ _service { sdl } }'
-  let res = await app.inject({ method: 'GET', url: `/graphql?query=${query}` })
-  t.same(JSON.parse(res.body), { data: { _service: { sdl: schema } } })
-
-  query = 'query { foo }'
-  res = await app.inject({ method: 'POST', url: '/graphql', body: { query } })
-  t.same(JSON.parse(res.body), { data: { foo: 'BAR' } })
 })
 
 test('entities resolver returns correct value for interface', async (t) => {
@@ -1350,9 +1241,9 @@ test('entities resolver returns correct value for interface', async (t) => {
     }
   }
 
-  const federationSchema = buildFederationSchema(schema)
+  const federationSchema = buildFederationSchema(gql(schema))
 
-  app.register(GQL, {
+  app.register(mercurius, {
     schema: federationSchema,
     resolvers
   })
@@ -1378,7 +1269,7 @@ test('entities resolver returns correct value for interface', async (t) => {
     url: `/graphql?query=${query}`
   })
 
-  t.same(JSON.parse(res.body), {
+  t.assert.deepStrictEqual(JSON.parse(res.body), {
     data: {
       _entities: [{
         __typename: 'BundleProduct',
@@ -1433,9 +1324,9 @@ test('entities resolver returns correct value for interface with async resolveTy
     }
   }
 
-  const federationSchema = buildFederationSchema(schema)
+  const federationSchema = buildFederationSchema(gql(schema))
 
-  app.register(GQL, {
+  app.register(mercurius, {
     schema: federationSchema,
     resolvers
   })
@@ -1461,7 +1352,7 @@ test('entities resolver returns correct value for interface with async resolveTy
     url: `/graphql?query=${query}`
   })
 
-  t.same(JSON.parse(res.body), {
+  t.assert.deepStrictEqual(JSON.parse(res.body), {
     data: {
       _entities: [{
         __typename: 'BundleProduct',
@@ -1516,9 +1407,9 @@ test('entities resolver returns correct value for interface with async resolveTy
     }
   }
 
-  const federationSchema = buildFederationSchema(schema)
+  const federationSchema = buildFederationSchema(gql(schema))
 
-  app.register(GQL, {
+  app.register(mercurius, {
     schema: federationSchema,
     resolvers
   })
@@ -1544,7 +1435,7 @@ test('entities resolver returns correct value for interface with async resolveTy
     url: `/graphql?query=${query}`
   })
 
-  t.same(JSON.parse(res.body), {
+  t.assert.deepStrictEqual(JSON.parse(res.body), {
     data: {
       _entities: [{
         __typename: 'BundleProduct',
@@ -1591,9 +1482,9 @@ test('entities resolver returns correct value for interface with async without d
     BundleProduct: {}
   }
 
-  const federationSchema = buildFederationSchema(schema)
+  const federationSchema = buildFederationSchema(gql(schema))
 
-  app.register(GQL, {
+  app.register(mercurius, {
     schema: federationSchema,
     resolvers
   })
@@ -1615,7 +1506,7 @@ test('entities resolver returns correct value for interface with async without d
     url: `/graphql?query=${query}`
   })
 
-  t.same(JSON.parse(res.body), {
+  t.assert.deepStrictEqual(JSON.parse(res.body), {
     data: {
       _entities: [{
         __typename: 'BundleProduct',
@@ -1682,9 +1573,9 @@ test('entities resolver returns correct value for interface using isTypeOf metho
     }
   }
 
-  const federationSchema = buildFederationSchema(schema)
+  const federationSchema = buildFederationSchema(gql(schema))
 
-  app.register(GQL, {
+  app.register(mercurius, {
     schema: federationSchema,
     resolvers
   })
@@ -1710,7 +1601,7 @@ test('entities resolver returns correct value for interface using isTypeOf metho
     url: `/graphql?query=${query}`
   })
 
-  t.same(JSON.parse(res.body), {
+  t.assert.deepStrictEqual(JSON.parse(res.body), {
     data: {
       _entities: [{
         __typename: 'BundleProduct',
@@ -1779,9 +1670,9 @@ test('entities resolver returns correct value for interface using isTypeOf as pr
     }
   }
 
-  const federationSchema = buildFederationSchema(schema)
+  const federationSchema = buildFederationSchema(gql(schema))
 
-  app.register(GQL, {
+  app.register(mercurius, {
     schema: federationSchema,
     resolvers
   })
@@ -1807,7 +1698,7 @@ test('entities resolver returns correct value for interface using isTypeOf as pr
     url: `/graphql?query=${query}`
   })
 
-  t.same(JSON.parse(res.body), {
+  t.assert.deepStrictEqual(JSON.parse(res.body), {
     data: {
       _entities: [{
         __typename: 'BundleProduct',
